cleaned up the code duplication
