+++ backend/src/utils.js
@@ -9,8 +9,27 @@ class Result {
   }
 }
 
+function promiseResult(promise) {
+  // Convert a promise to an always-resolving promise of Result type.
+  return new Promise((resolve) => {
+    const payload = new Result();
+    promise
+      .then((result) => {
+        payload.value = result;
+      })
+      .catch((error) => {
+        payload.error = error;
+      })
+      .then(() => {
+        resolve(payload);
+      });
+  });
+}
+
 function delayFor(milliseconds) {
   return new Promise((resolve) => setTimeout(resolve, milliseconds));
 }
 
 exports.Result = Result;
+exports.promiseResult = promiseResult;
+exports.delayFor = delayFor;
